sperate logic from view in navbar
